% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_data}
\alias{build_data}
\title{Create do wrapper function with source data}
\usage{
build_data(funcname)
}
\value{
do wrapper function
}
\description{
Create do wrapper function with source data
}

