% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_glm}
\alias{build_glm}
\title{glm wrapper with do}
\usage{
build_glm(df, ..., keep.source = FALSE, augment = FALSE)
}
\value{
deta frame which has glm model
}
\description{
glm wrapper with do
}

