% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_kmeans.kv}
\alias{build_kmeans.kv}
\title{kmeans wrapper with do}
\usage{
build_kmeans.kv(df, subject, key, value, centers = 3, iter.max = 10,
  nstart = 1, algorithm = "Hartigan-Wong", trace = FALSE,
  keep.source = FALSE, seed = 0, augment = F, fun.aggregate = mean,
  fill = 0)
}
\arguments{
\item{centers}{Set an integer number to decide how many clusters (groups) to build.}

\item{keep.source}{It will make .source.data column to preserve source data.}

\item{seed}{This is random seed. You can change the result if you change this number.}
}
\value{
deta frame which has kmeans model
}
\description{
kmeans wrapper with do
}

