% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_kmeans.variables}
\alias{build_kmeans.variables}
\title{kmeans wrapper with do with variable columns as input}
\usage{
build_kmeans.variables(df, ..., centers = 3, iter.max = 10, nstart = 1,
  algorithm = "Hartigan-Wong", trace = FALSE, keep.source = FALSE,
  seed = 0, augment = FALSE)
}
\description{
kmeans wrapper with do with variable columns as input
}

