% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_t.test}
\alias{build_t.test}
\title{t.test wrapper with do}
\usage{
build_t.test(df, ..., keep.source = FALSE, augment = F)
}
\value{
deta frame which has t.test model
}
\description{
t.test wrapper with do
}

