% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{build_var.test}
\alias{build_var.test}
\title{var.test wrapper with do}
\usage{
build_var.test(df, ..., keep.source = FALSE, augment = F)
}
\value{
deta frame which has var.test model
}
\description{
var.test wrapper with do
}

