% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{calc_cor_cat}
\alias{calc_cor_cat}
\title{Calculate correlation among categories and output the result in tidy format}
\usage{
calc_cor_cat(df, category_col, dimention_col, value_col,
  use = "pairwise.complete.obs", method = "pearson", fun.aggregate = mean)
}
\arguments{
\item{df}{data frame in tidy format}

\item{category_col}{A name of column which contains values of categories}

\item{dimention_col}{A name of column which has dimentions of vectors}

\item{value_col}{A name of column which has values}

\item{use}{Operation type for dealing with missing values. This can be one of "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs"}

\item{method}{Method of calculation. This can be one of "pearson", "kendall", or "spearman".}
}
\value{
correlations between pairs of categories
}
\description{
Calculate correlation among categories and output the result in tidy format
}

