% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{calc_doc_sim}
\alias{calc_doc_sim}
\title{Calculate similarity of each pair of groups.}
\usage{
calc_doc_sim(df, group, dimention, value, distinct = FALSE, diag = FALSE)
}
\arguments{
\item{df}{data frame in tidy format}

\item{group}{A column you want to calculate the correlations for.}

\item{value}{A column for the values you want to use to calculate the correlations.}

\item{distinct}{The returned pair should be duplicated in swapped order or not.
TRUE makes it easy to filter group names.}

\item{diag}{If similarity between itself should be returned or not.}

\item{dimension}{A column you want to use as a dimension to calculate the correlations.}

\item{method}{Type of calculation. https://cran.r-project.org/web/packages/proxy/vignettes/overview.pdf}
}
\description{
Calculate similarity of each pair of groups.
}

