% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{calc_tf_}
\alias{calc_tf_}
\title{Caluculate tf for standard evaluation}
\usage{
calc_tf_(tbl, document_col, term_col, weight = "ratio", k = 0.5)
}
\description{
Caluculate tf for standard evaluation
}

