% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{calc_tfidf}
\alias{calc_tfidf}
\title{Caluculate tfidf}
\usage{
calc_tfidf(tbl, document, term, idf_log_scale = log, smooth_idf = FALSE,
  tf_weight = "ratio", tf_k = 0.5)
}
\description{
Caluculate tfidf
}

