% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_package_versions}
\alias{check_package_versions}
\title{Check package update from cran and requiredRpackage json}
\usage{
check_package_versions(json_path, download = NULL, overwrite = FALSE)
}
\description{
Check package update from cran and requiredRpackage json
}

