% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{clean_data_frame}
\alias{clean_data_frame}
\title{This function can clean the given data frame. It actually does 
1) split a column with a data.frame vector into seprate columns 
2) repair column names such as columns with NA for column names, 
or duplicate column names.}
\usage{
clean_data_frame(x)
}
\arguments{
\item{x}{data frame}
}
\value{
cleaned data frame
}
\description{
This function can clean the given data frame. It actually does 
1) split a column with a data.frame vector into seprate columns 
2) repair column names such as columns with NA for column names, 
or duplicate column names.
}

