% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_cmdscale}
\alias{do_cmdscale}
\title{Calculate svd from tidy format. This can be used to calculate coordinations by reducing dimensionality.}
\usage{
do_cmdscale(df, pair_name1, pair_name2, value, k = 2, eig = FALSE,
  fun.aggregate = mean, fill = 0)
}
\arguments{
\item{df}{Data frame which has group and dimension}
}
\value{
Tidy format of data frame.
}
\description{
Calculate svd from tidy format. This can be used to calculate coordinations by reducing dimensionality.
}

