% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_cor.cols}
\alias{do_cor.cols}
\title{Calculate correlation among columns and output the correlation of each pair}
\usage{
do_cor.cols(df, ..., use = "pairwise.complete.obs", method = "pearson",
  distinct = FALSE, diag = FALSE)
}
\arguments{
\item{df}{data frame in tidy format}

\item{...}{Arguments to select columns to calculate correlation.}

\item{use}{Operation type for dealing with missing values. This can be one of "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs"}

\item{method}{Method of calculation. This can be one of "pearson", "kendall", or "spearman".}
}
\value{
correlations between pairs of columns
}
\description{
Calculate correlation among columns and output the correlation of each pair
}

