% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_cor.kv}
\alias{do_cor.kv}
\title{Calculate correlation among groups and output the correlation of each pair}
\usage{
do_cor.kv(df, subject, key, value, use = "pairwise.complete.obs",
  method = "pearson", distinct = FALSE, diag = FALSE, fill = 0,
  fun.aggregate = mean)
}
\arguments{
\item{df}{data frame in tidy format}

\item{value}{A column for the values you want to use to calculate the correlations.}

\item{use}{Operation type for dealing with missing values. This can be one of "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs"}

\item{method}{Method of calculation. This can be one of "pearson", "kendall", or "spearman".}

\item{fun.aggregate}{Set an aggregate function when there are multiple entries for the key column per each category.}

\item{group}{A column you want to calculate the correlations for.}

\item{dimension}{A column you want to use as a dimension to calculate the correlations.}
}
\value{
correlations between pairs of groups
}
\description{
Calculate correlation among groups and output the correlation of each pair
}

