% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{do_dist.kv}
\alias{do_dist.kv}
\title{Calculate distance of each pair of groups.}
\usage{
do_dist.kv(df, subject, key, value, fill = 0, fun.aggregate = mean,
  distinct = FALSE, diag = FALSE, method = "euclidean", p = 2)
}
\arguments{
\item{df}{data frame in tidy format}

\item{value}{A column for the values you want to use to calculate the correlations.}

\item{distinct}{The returned pair should be duplicated in swapped order or not.
TRUE makes it easy to filter group names.}

\item{diag}{If similarity between itself should be returned or not.}

\item{method}{Type of calculation. https://cran.r-project.org/web/packages/proxy/vignettes/overview.pdf}

\item{group}{A column you want to calculate the correlations for.}

\item{dimension}{A column you want to use as a dimension to calculate the correlations.}
}
\description{
Calculate distance of each pair of groups.
}

