% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{do_ngram}
\alias{do_ngram}
\title{Generate ngram in groups.}
\usage{
do_ngram(df, token, sentence, document, maxn = 2, sep = "_")
}
\arguments{
\item{df}{Data frame which has tokens.}

\item{token}{Column name of token data.}

\item{n}{How many tokens should be together as new tokens. This should be numeric vector.}
}
\description{
Generate ngram in groups.
}

