% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_svd.kv}
\alias{do_svd.kv}
\title{Calculate svd from tidy format. This can be used to calculate coordinations by reducing dimensionality.}
\usage{
do_svd.kv(df, subject, key, value, type = "group", fill = 0,
  fun.aggregate = mean, n_component = 3, centering = TRUE,
  output = "long")
}
\arguments{
\item{df}{Data frame which has group and dimension}

\item{value}{Column to be regarded as values}

\item{type}{"group" to see the coordinations in reduced dimension.
"dimension" to see the direction of new axes from original ones.
"variance" to see how much the data is distributed in the direction of new axes.}

\item{fill}{Value to fill where value doesn't exist.}

\item{fun.aggregate}{Value to fill where value doesn't exist.}

\item{n_component}{Number of dimensions to return.}

\item{group}{Column to be regarded as groups}

\item{dimension}{Column to be regarded as original dimensions}
}
\value{
Tidy format of data frame.
}
\description{
Calculate svd from tidy format. This can be used to calculate coordinations by reducing dimensionality.
}

