% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_svd_var}
\alias{do_svd_var}
\title{Calculate svd from spread format. This can be used to calculate coordinations by reducing dimensionality.}
\usage{
do_svd_var(df, ..., label = NULL, fill = 0, n_component = 3,
  centering = TRUE, type = "group")
}
\description{
Calculate svd from spread format. This can be used to calculate coordinations by reducing dimensionality.
}

