% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{do_svd_var_}
\alias{do_svd_var_}
\title{SE version of do_svd_var}
\usage{
do_svd_var_(df, columns, label_col = NULL, n_component = 3,
  centering = TRUE, type = "group")
}
\description{
SE version of do_svd_var
}

