% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{do_tfidf}
\alias{do_tfidf}
\title{Calculate tfidf, which shows how much particular the token is in a group.}
\usage{
do_tfidf(df, group, term, idf_log_scale = log, tf_weight = "raw",
  tf_k = 0.5, norm = "l2")
}
\arguments{
\item{df}{Data frame which has columns of groups and their terms}

\item{group}{Column of group names}

\item{term}{Column of terms}

\item{idf_log_scale}{Function to scale IDF. It might be worth trying log2 or log10.
log10 strongly suppress the increase of idf values and log2 does it more weakly.}
}
\description{
Calculate tfidf, which shows how much particular the token is in a group.
}

