% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{do_var.test}
\alias{do_var.test}
\title{var.test wrapper with do}
\usage{
do_var.test(df, ..., keep.source = FALSE)
}
\value{
deta frame which has var.test model
}
\description{
var.test wrapper with do
}

