% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Get sentiments of words}
\usage{
get_sentiment(words, lexicon = "bing")
}
\arguments{
\item{words}{Vector of words to check sentiment.}

\item{lexicon}{Type of sentiment. One of "nrc" "bing" "AFINN".}
}
\value{
Vector of sentiment.
}
\description{
Get sentiments of words
}

