% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Get vector of stopwords}
\usage{
get_stopwords(lexicon = "snowball")
}
\arguments{
\item{lexicon}{Type of stopwords. One of "snowball", "onix" and "SMART".}
}
\value{
vector of stop word.
}
\description{
Get vector of stopwords
}

