% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocation.R
\name{ip_to_country}
\alias{ip_to_country}
\title{Convert IP address to country name}
\usage{
ip_to_country(ip)
}
\arguments{
\item{ip}{IP address}
}
\value{
Country name
}
\description{
Convert IP address to country name
}

