% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{is_digit}
\alias{is_digit}
\title{Check if the word is digits.}
\usage{
is_digit(word)
}
\arguments{
\item{word}{Character to be checked if it's digits.}
}
\value{
Logical vector if the word is digits or not.
}
\description{
Check if the word is digits.
}

