% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{is_stopword}
\alias{is_stopword}
\title{Check if the token is in stopwords.}
\usage{
is_stopword(token, lexicon = "snowball")
}
\arguments{
\item{token}{Character to be checked if it's stopword.}

\item{lexicon}{Type of stopwords. One of "snowball", "onix" and "SMART".}
}
\value{
Logical vector if the token is in stopwords or not.
}
\description{
Check if the token is in stopwords.
}

