% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise_dist}
\alias{pairwise_dist}
\title{Calc cosine similarity}
\usage{
pairwise_dist(tbl, group, dimention, value, upper = FALSE, diag = FALSE,
  method = "Euclidean")
}
\description{
Calc cosine similarity
}

