% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise_sim}
\alias{pairwise_sim}
\title{Calc similarity similarity}
\usage{
pairwise_sim(tbl, group, dimention, value, upper = FALSE, diag = FALSE,
  method = "correlation")
}
\description{
Calc similarity similarity
}

