% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_wrapper.R
\name{reduce_dimension}
\alias{reduce_dimension}
\title{Compress dimension using svd algorithm}
\usage{
reduce_dimension(df, group, dimension, value, type = "group", fill = 0,
  fun.aggregate = mean, dim = NULL)
}
\arguments{
\item{group}{Column to be regarded as groups}

\item{dimension}{Column to be regarded as original dimensions}

\item{value}{Column to be regarded as values}

\item{value}{Column to be regarded as values
Compress dimension}
}
\description{
Compress dimension using svd algorithm
}

