% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{stem_word}
\alias{stem_word}
\title{Stem word so that words which are the same kind of word become the same charactor}
\usage{
stem_word(...)
}
\arguments{
\item{language}{This can be "porter" or a kind of languages which use alphabet like "english" or "french".}
}
\description{
Stem word so that words which are the same kind of word become the same charactor
}

