% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to_matrix}
\alias{to_matrix}
\title{as.matrix from select argument or cast by three columns}
\usage{
to_matrix(df, select_dots, by_col = NULL, key_col = NULL,
  value_col = NULL, fill = 0, fun.aggregate = mean)
}
\description{
as.matrix from select argument or cast by three columns
}

