% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{upper_gather}
\alias{upper_gather}
\title{Gather only right upper half of matrix - where row_num > col_num}
\usage{
upper_gather(mat, names = NULL, diag = NULL, cnames = c("Var1", "Var2",
  "value"))
}
\description{
Gather only right upper half of matrix - where row_num > col_num
}

