% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{calc_sim}
\alias{calc_sim}
\title{Calc similarity similarity}
\usage{
calc_sim(df, group, dimention, value, upper = FALSE, diag = FALSE,
  method = "correlation")
}
\description{
Calc similarity similarity
}

