% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{calc_tf}
\alias{calc_tf}
\alias{calc_tf_}
\title{Calculate term frequency}
\usage{
calc_tf(df, group, term, ...)

calc_tf_(df, group_col, term_col, weight = "ratio", k = 0.5)
}
\arguments{
\item{df}{Data frame}

\item{group}{Column to be considered as a group id}

\item{term}{Column to be considered as term}

\item{weight}{Type of weight calculation.
"ratio" is default and it's count/(total number of terms in the group).
This can be "raw_frequency", "binary", "log_normalization" and "k_normalization"
"raw_frequency" is the count of the term in the group.
"binary" is logic if the term is in the group or not.
"log_normalization" is logic if the term is in the group or not.}

\item{term}{Column to be considered as term}
}
\value{
Data frame with group, term and .tf column
}
\description{
Calculate term frequency
}

