% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{do_ngram}
\alias{do_ngram}
\title{Generate ngram in groups.}
\usage{
do_ngram(df, token, n = 1:2, skip = 0)
}
\arguments{
\item{df}{Data frame which has tokens.}

\item{token}{Column name of token data.}

\item{n}{How many tokens should be together as new tokens. This should be numeric vector.}

\item{skip}{How many tokens can be skipped when connecting them.}
}
\description{
Generate ngram in groups.
}

