% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operation.R
\name{do_tokenize}
\alias{do_tokenize}
\title{Tokenize text and unnest}
\usage{
do_tokenize(df, input, output = .token, ...)
}
\arguments{
\item{df}{Data frame}

\item{input}{Input column name}

\item{output}{Output column name}
}
\value{
Data frame with tokenized column
}
\description{
Tokenize text and unnest
}

